pyproject.toml
[tool.poetry]
name = "memoriam"
version = "0.66.3"
description = "Knowledge management platform"
authors = ["Protojour dev team <devs@protojour.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
aiocontextvars = "^0.2.2"
ariadne = "^0.20.1"
brotli = "^1.1.0"
case-converter = "1.0.2"
cashews = { version = "5.0.0", extras = ["redis"] }
click = "^8.1.7"
graphql-core = "3.2.1"
hiredis = "^2.3.2"
httpx = { version = "0.23.3", extras = ["http2"] }
httpx-ws = "^0.3.1"
jinja2 = "^3.1.4"
jsonschema-rs = "^0.18.0"
migrado = "^0.6.11"
minio = "^7.2.7"
python-benedict = "^0.33.0"
pyyaml = "^6.0.1"
trustme = "^1.1.0"
sanic = "^22.12.0"
sanic-ext = "^22.12.0"
sentry-sdk = "^2.3.0"
ujson = "^5.10.0"
uvloop = "^0.19.0"

[tool.poetry.group.docs.dependencies]
furo = "^2023.9.10"
myst-parser = "^2.0.0"
sphinx = "^7.2.6"

[tool.poetry.group.dev.dependencies]
docker = "^6.1.3"
pytest = "^7.1.2"
pytest-asyncio = "^0.21.1"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.1.0"
pytest-env = "^1.0.1"
pytest-httpx = "^0.21.1"
pytest-parallel = "^0.1.1"
pytest-repeat = "^0.9.1"
ruff = "^0.4.7"
sanic-testing = "^22.12.0"

[tool.poetry.scripts]
memoriam = "memoriam.cli:memoriam"
migrado = "migrado:migrado"

[tool.ruff]
lint.select = ["E", "F"]
lint.ignore = ["E501"]

[build-system]
requires = ["poetry>=1.2.0"]
build-backend = "poetry.core.masonry.api"
